# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.TmdbLookupResponse:
    properties:
      mediaType:
        enum:
        - movie
        - tv
        type: string
      poster:
        type: string
      title:
        type: string
      tmdbId:
        type: string
      year:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Standalone BitTorrent streaming engine for Potok: torrent metadata, direct streaming, HLS, subtitles, thumbnails, and an optional management web UI.
    The plugin/streaming routes are open (players hit them without credentials); the /api/manage/* contour is mounted only with TORRENTGO_ENABLE_WEBUI=true and protected by HTTP Basic auth. Errors are text/plain (via http.Error) unless noted otherwise.
  title: Potok TorrentGo API
  version: "2.0"
paths:
...
  /api/manage/tmdb:
    get:
      description: Resolves a TMDB id to a title + poster URL + year for the Add-torrent
        dialog. Proxies TMDB (v3 api_key from TMDB_API_KEY) so no key is exposed to
        the browser. Returns 501 if no key is set — manual title/poster entry still
        works without it.
      operationId: tmdbLookup
      parameters:
      - description: TMDB id
        in: query
        name: id
        required: true
        type: string
      - description: movie (default) | tv
        enum:
        - movie
        - tv
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TmdbLookupResponse'
        "400":
          description: id required
          schema:
            type: string
        "401":
          description: unauthorized (BasicAuth)
          schema:
            type: string
        "501":
          description: TMDB not configured (set TMDB_API_KEY)
          schema:
            type: string
        "502":
          description: TMDB request/lookup failed
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Look up title/poster by TMDB id
      tags:
      - Management
      x-scalar-ignore: true
...
  /api/torrents/{hash}/files/{fileIndex}/thumbnail:
    get:
      description: Seeks in-process to `time` (rounded to a 5-second bucket) and returns
        a 160x90 JPEG. Results are LRU-cached; concurrent requests for the same bucket
        are coalesced.
      operationId: getThumbnail
      parameters:
      - description: Infohash (40-char hex)
        in: path
        name: hash
        required: true
        type: string
      - description: File index in the torrent (1-based)
        in: path
        name: fileIndex
        required: true
        type: integer
      - description: Timestamp in seconds (rounded to the nearest 5s)
        in: query
        name: time
        type: number
      produces:
      - image/jpeg
      responses:
        "200":
          description: JPEG image (160x90)
          schema:
            type: string
        "400":
          description: invalid hash or file index
          schema:
            type: string
        "404":
          description: torrent not active
          schema:
            type: string
        "500":
          description: thumbnail extraction failed
          schema:
            type: string
      security: []
      summary: Extract a JPEG thumbnail at a timestamp
      tags:
      - Media
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"